backend/pyproject.toml
[project]
name = "user-crud-oauth2-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.15,<4.0",
  "django-oauth-toolkit>=2.4,<3.0",
  "drf-spectacular>=0.27,<1.0",
  "django-cors-headers>=4.3,<5.0",
  "psycopg[binary]>=3.1,<4.0",
  "python-dotenv>=1.0,<2.0",
  "pytest>=8.0,<9.0",
  "pytest-django>=4.8,<5.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
addopts = "-q"
